Add files via upload
Project Brief
The issue of retaining one's employees happy and satisfied is a challenge.
If an employee that you have invested so much time and money leaves for other companies, then this would mean that you would have to spend even more time and money to hire somebody else. 
Making use of data science and predictive modeling capabilities,  we can predict employee turnover rate and it will save the company from loss.